TAP test to demonstrate remote fault injector interface
The test verifies that commit on master waits until a synchronous
standby has flush WAL up to or greater than commit LSN.  A new fault
point 'standby_flush' is defined in WAL receiver.  When it is enabled
with type 'skip', the WAL receiver responds with stale flush LSN value.

The test invokes faultinjector interface on master to inject the fault
on standby.  The faultinjector interface can connect to a remote server
based on hostname and port.

Note: the test depends on non-core Perl module IPC::Run.  The
IPC::Run::start interface is used to start a psql session in background.
If this is not acceptable, I am sure there are in-core options to
achieve background subprocesses, however, with my limited Perl skills
(as evident from the test), this seems to serve the purpose as a first
attempt.