Start WAL receiver when it is found not running
Postmaster now starts WAL receiver as soon as it is found not running
from ServerLoop.  This helps to resume streaming replication sooner
when a temporary network disruption causes WAL receiver process to
exit.

As a consequence, the race condition addressed in e5d494d78cf is
eliminated.  Postmaster may start WAL receiver in states that allow a
standby to operate, except PM_STARTUP.  It is not possible to
distinguish whether the postmaster is operating as master or standby in
PM_STARTUP state.

Postmaster attempts to start WAL receiver as long as a promote
request is not received and the state permits.

Co-authored-by: Asim R P <apraveen@pivotal.io>