Start WAL receiver before startup process replays existing WAL
If WAL receiver is started only after startup process finishes replaying
WAL already available in pg_wal, synchornous replication is impacted
adversly.  Consider a temporary network outage causing streaming
replication connection to break.  This leads to exit of WAL receiver
process.  If the startup process has fallen behind, it may take a long
time to finish replaying WAL and then start walreceiver again to
re-establish streaming replication.  Commits on master will have to wait
all this while for the standby to flush WAL upto commit LSN.

This experience can be alleviated if replication connection is
re-established as soon as it is found to be disconnected.  The patch
attempts to make this happen by starting WAL receiver from postmaster
ServerLoop as well as startup process, even before it WAL replay begins.

The start point to request streaming from is remembered in pg_control.
Before creating a new WAL segment file, WAL receiver records the new WAL
segment number in pg_control.  If the WAL receiver process exits and
must restart, the recorded segment number is used to generate a start
point, that is the first offset in the segment file to re-establish
streaming replication.

Alternatives we thought of (but did not implement) for persisting the
starting point: (1) postgresql.auto.conf file, similar to how
primary_conninfo is remembered.  This option requires creating a new GUC
that represents the starting point.  Start point is never set by a user,
so using a GUC to represent it does not seem appropriate.  (2) introduce
a new flat file.  This incurs the overhead to maintain an additional
flat file.

Co-authored-by: Asim R P <apraveen@pivotal.io>