Test that replay lag does not affect syncrep
Synchronous replication, when synchronous_commit is set to remote
flush, should not be affected by replay lag.  This test disconnects
WAL replication connection after inducing replay lag.  The objective
is to verify that replication connection is reestablished and standby
continues to stream and write WAL, in-spite of the replay lag.